Update bellerophon.kicad_sch
Added surface mount resistors  (1206 form) and surface mount Mosfet see links attached to the components for the ones to buy off RS components.

Added a continuity check, this circuit is always live but draws a tiny current, it can be used to check the eMatch is present. A high signal of  ~2.4-3V (via a voltage divider) depending on power source can be detected (should be above the logic high level of a teensy). and when 0V is detected then the pyro is either not connected OR it has been fired. DO NOT CONNECT A VOLTAGE SOURCE HIGHER THAN 10 VOLTS - This will result in a signal of >3.3 V and could damage teensy pins.

Brown out capacitor wont be useful here I don't think. I suspect using a brown out capacitor on the teensy side to keep a constant voltage to the teensy would be much more useful. The pyro circuit should only be active for a short time as per teensy code.